подправил тип значения

+++ src/obj.hpp
@@ -391,8 +391,8 @@ public:
 	float show_remort_req();
 	float show_mort_req();
 	float show_koef_obj();
-	unsigned get_ilevel() const;	///< разные системы расчета привлекательности предмета
-	void set_ilevel(unsigned ilvl);
+	float get_ilevel() const;	///< разные системы расчета привлекательности предмета
+	void set_ilevel(float ilvl);
 	auto get_rnum() const { return m_rnum; }
 	void set_rnum(const obj_rnum _);
 	auto get_vnum() const { return m_vnum; }
@@ -461,7 +461,7 @@ private:
 	int m_rent_on;	///< стоимость ренты, если надета
 	int m_rent_off;	///< стоимость ренты, если в инве
 
-	unsigned m_ilevel;	///< расчетный уровень шмотки, не сохраняется
+	float m_ilevel;	///< расчетный уровень шмотки, не сохраняется
 	obj_vnum m_rnum;	///< Where in data-base
 
 	std::unordered_set<VNumChangeObserver::shared_ptr> m_vnum_change_observers;
